requirements.txt
Django>=5.0,<5.1
gunicorn
channels
numpy==1.26.3
django-allauth[socialaccount]
opencv-python
imutils
cmake
dlib-19.24.99-cp312-cp312-win_amd64.whl
scipy
playsound==1.2.2
camera
decouple